pyproject.toml
[tool.poetry]
name = "ad-looper"
version = "0.1.0"
description = ""
authors = ["Demianight <komarovdemian@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
aiofiles = "^24.1.0"
aiosqlite = "^0.20.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pytest-asyncio = "^0.23.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
isort = "^5.13.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "manage:start"
startapp = "manage:startapp"
tests = "manage:tests"
pre_commit = "manage:pre_commit"
makemigrations = "manage:makemigrations"
migrate = "manage:migrate"
